[dining event] event point bugs fixed, calendar fixed, quit added, only miss admin side *all +edit

+++ application/controllers/EventInfo.php
@@ -13,11 +13,14 @@ class EventInfo extends CI_Controller
         parent::__construct();
         $this->load->helper('url');
         $this->load->library('session');
+        $this->load->library('javascript');
         $this->load->database();
         $this->load->model('GetEventInfo_model');
         $this->load->model('CreateEvent_model');
         $this->load->model('EditPoint_model');
         $this->load->model('DeleteEvent_model');
+        $this->load->model('JoinEvent_model');
+        $this->load->model('GetMemberInfo_model');
         $this->load->library('form_validation');
 
         $this->load->model('SearchEvent_model','');
@@ -60,6 +63,29 @@ class EventInfo extends CI_Controller
             $this->data['eventInfo']=$this->GetEventInfo_model->get_particular_event($id);
 
             $this->load->view('showparticularevent_view',$this->data);
+
+            $this->load->view('template/footer');
+
+        } else {
+            //If no session, redirect to login page
+            redirect('Welcome', 'refresh');
+        }
+
+
+    }
+
+    public function viewjoint($id = '') {
+
+        if($this->session->userdata('logged_in')){
+
+            $data['title'] = 'View Joint Dining Event Information';
+
+            $this->load->view('template/navigation_member', $data);
+            $this->load->view('template/header', $data);
+
+            $this->data['eventInfo']=$this->GetEventInfo_model->get_particular_event($id);
+
+            $this->load->view('showparticularjointevent_view',$this->data);
 
             $this->load->view('template/footer');
 
@@ -166,19 +192,19 @@ class EventInfo extends CI_Controller
             array(
                 'field' => 'eventMinParti',
                 'label' => 'Min. Participant',
-                'rules' => 'trim|required|integer',
+                'rules' => 'trim|required|integer|greater_than[1]|max_length[2]',
             ),
 
             array(
                 'field' => 'eventMaxParti',
                 'label' => 'Max. Participant',
-                'rules' => 'trim|required|integer',
+                'rules' => 'trim|required|integer|greater_than[1]|max_length[2]',
             ),
 
             array(
                 'field' => 'eventEstFee',
                 'label' => 'Estimated Fee',
-                'rules' => 'trim|required|decimal',
+                'rules' => 'trim|required|greater_than[0]|callback_decimal_check',
             ),
 
             array(
@@ -212,6 +238,8 @@ class EventInfo extends CI_Controller
 
             $session_data = $this->session->userdata('logged_in');
             $session_memberName = $session_data['memberName'];
+            $session_memberID = $session_data['memberID'];
+
 
             $eventdata = array(
                 'eventCreatorName' =>  $session_memberName,
@@ -228,11 +256,20 @@ class EventInfo extends CI_Controller
                 'eventAddress' => $this->security->xss_clean($this->input->post('eventAddress')),
             );
 
-            $pointvalue = $eventdata['eventMemberPoint'];
 
-            $this->CreateEvent_model->form_insert($eventdata);
+            $eid = $this->CreateEvent_model->form_insert($eventdata);
 
-            $this->EditPoint_model->event_addpoint($session_memberName,$pointvalue);
+            $pointvalue = (int)$this->GetEventInfo_model->get_eventpoint($eid);
+
+            $nowpoint = (int)$this->GetMemberInfo_model->get_memberpoint($session_memberID);
+
+            $addto = $pointvalue+$nowpoint;
+
+
+            $this->JoinEvent_model->join_event($eid,$session_memberID);
+            $this->EditPoint_model->event_addpoint($session_memberName,$addto);
+
+
 
             $data['title'] = 'Create Dining Event';
             $this->load->view('template/navigation_member');
@@ -241,6 +278,17 @@ class EventInfo extends CI_Controller
             $this->load->view('template/footer');
         }
 
+
+    }
+
+    function decimal_check($input){
+
+            if(preg_match('/^\d+(\.\d{2})$/', $input)) {
+                return true;
+            }else{
+                $this->form_validation->set_message('decimal_check', 'The %s field is in wrong format. It should be in 2 d.p');
+                return false;
+            }
 
     }
 
@@ -253,7 +301,7 @@ class EventInfo extends CI_Controller
         }
         else
         {
-            $this->form_validation->set_message('datetime_check', 'The %s field is in wrong format.');
+            $this->form_validation->set_message('datetime_check', 'The %s field is in wrong format. It should be in yyyy-MM-dd hh:mm:ss format');
             return false;
         }
     }
@@ -264,16 +312,39 @@ class EventInfo extends CI_Controller
 
             $session_data = $this->session->userdata('logged_in');
             $data['memberName'] = $session_data['memberName'];
+            $name = $session_data['memberName'];
             $data['title'] = 'Delete Dining Event';
 
 
             $this->load->view('template/navigation_member', $data);
             $this->load->view('template/header', $data);
 
-            $this->load->model('DeleteEvent_model');
-            $this->DeleteEvent_model->delete_event($id);
+            $eventcreatorname = $this->GetEventInfo_model->get_eventcreatorname($id);
 
-            $this->load->view('sucessful_deleteevent');
+            if ($eventcreatorname == $name){
+
+
+                $pointvalue = (int)$this->GetEventInfo_model->get_eventpoint($id);
+
+                $mid = $this->GetMemberInfo_model->get_memberID($name);
+
+                $nowpoint = (int)$this->GetMemberInfo_model->get_memberpoint($mid);
+
+                $minusto = $nowpoint-$pointvalue;
+                $this->EditPoint_model->event_minuspoint($name,$minusto);
+
+                $this->load->model('DeleteEvent_model');
+                $this->DeleteEvent_model->delete_event($id);
+                $this->DeleteEvent_model->delete_jointevent($id);
+
+                $this->load->view('sucessful_deleteevent');
+
+            } else {
+
+                $this->load->view('error_notcreatorcannotdeleteevent');
+
+            }
+
 
             $this->load->view('template/footer');
 
@@ -347,7 +418,7 @@ class EventInfo extends CI_Controller
 
     }
 
-    function check_database($eventid)
+    function check_database()
     {
         //Field validation succeeded.  Validate against database
         $eventID = $this->security->xss_clean($this->input->post('eventID'));
@@ -364,6 +435,156 @@ class EventInfo extends CI_Controller
             $this->form_validation->set_message('check_database', 'Invalid Event ID.');
             return false;
         }
+
+    }
+
+    public function joinevent ($id){
+
+        if($this->session->userdata('logged_in')){
+
+            $session_data = $this->session->userdata('logged_in');
+            $data['memberID'] = $session_data['memberID'];
+            $data['memberName'] = $session_data['memberName'];
+            $mid = $session_data['memberID'];
+            $mname = $session_data['memberName'];
+
+            $data['title'] = 'Join Dining Event';
+
+
+            $this->load->view('template/navigation_member', $data);
+            $this->load->view('template/header', $data);
+
+
+
+            $maxparti = $this->GetEventInfo_model->get_eventmax($id);
+            $nowparti = $this->JoinEvent_model->get_eventpartinum($id);
+
+
+            if($nowparti < $maxparti) {
+
+                $result = $this->JoinEvent_model->join_event($id,$mid);
+
+                if($result){
+
+                    $pointvalue = (int)$this->GetEventInfo_model->get_eventpoint($id);
+
+                    $nowpoint = (int)$this->GetMemberInfo_model->get_memberpoint($mid);
+
+                    $addto = $pointvalue+$nowpoint;
+
+                    $this->JoinEvent_model->join_event($id,$mid);
+                    $this->EditPoint_model->event_addpoint($mname,$addto);
+
+                    $this->load->view('sucessful_joinevent');
+                } else {
+
+                    $this->load->view('error_jointeventalready');
+                }
+
+
+            } else {
+
+                $this->load->view('error_eventreachmaxalready');
+
+            }
+
+
+
+            $this->load->view('template/footer');
+
+        } else {
+            //If no session, redirect to login page
+            redirect('Welcome', 'refresh');
+        }
+
+
+    }
+
+    public function joint() {
+
+        if($this->session->userdata('logged_in')){
+
+            $session_data = $this->session->userdata('logged_in');
+            $data['memberName'] = $session_data['memberName'];
+            $data['memberID'] = $session_data['memberID'];
+            $mid = $session_data['memberID'];
+            $data['title'] = 'View Joint Event';
+
+            $this->load->view('template/navigation_member', $data);
+            $this->load->view('template/header', $data);
+
+            $this->data['eventInfo']=$this->GetEventInfo_model->get_jointevent($mid);
+
+            if (empty($this->data['eventInfo'])){
+
+                $this->load->view('resultnotfound_view');
+
+            } else{
+
+                $this->load->view('showalljointevent_view',$this->data);
+
+            }
+
+            $this->load->view('template/footer');
+
+        } else {
+            //If no session, redirect to login page
+            redirect('Welcome', 'refresh');
+        }
+
+    }
+
+    public function quit($eid=""){
+
+        if($this->session->userdata('logged_in')){
+
+            $session_data = $this->session->userdata('logged_in');
+            $data['memberName'] = $session_data['memberName'];
+            $name = $session_data['memberName'];
+            $mid = $session_data['memberID'];
+            $data['title'] = 'Quit Dining Event';
+
+
+            $this->load->view('template/navigation_member', $data);
+            $this->load->view('template/header', $data);
+
+            $isjoint = $this->JoinEvent_model->check_eventjoint($eid,$mid);
+
+            if ($isjoint == true){
+
+
+                $pointvalue = (int)$this->GetEventInfo_model->get_eventpoint($eid);
+
+//                $mid = $this->GetMemberInfo_model->get_memberID($name);
+
+                $nowpoint = (int)$this->GetMemberInfo_model->get_memberpoint($mid);
+
+                $minusto = $nowpoint-$pointvalue;
+                $this->EditPoint_model->event_minuspoint($name,$minusto);
+
+                $this->load->model('DeleteEvent_model');
+                $this->DeleteEvent_model->delete_jointevent($eid);
+
+                $this->load->view('sucessful_quitevent');
+
+            } else {
+
+                $this->load->view('error_notjointcannotquitevent');
+
+            }
+
+
+            $this->load->view('template/footer');
+
+
+        } else {
+            //If no session, redirect to login page
+            redirect('Welcome', 'refresh');
+        }
+
+
+
+
     }
